RAG ChatBot/requirements.txt
langchain
langchain_community
PyPDF2
python-dotenv
streamlit
Cohere
faiss-cpu
altair
tiktoken
huggingface-hub
InstructorEmbedding
sentence-transformers
pydantic
